Move form constructor to demo
